+++ src/java/com/smi/travel/datalayer/report/model/OtherAgentCommissionSummaryReport.java
@@ -5,6 +5,8 @@
  */
 package com.smi.travel.datalayer.report.model;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author Kanokporn
@@ -17,7 +19,7 @@ public class OtherAgentCommissionSummaryReport {
     private String code;
     private String name;
     private int countbook;
-    private int commission;
+    private BigDecimal commission;
 
     public String getSystemdate() {
         return systemdate;
@@ -75,11 +77,11 @@ public class OtherAgentCommissionSummaryReport {
         this.countbook = countbook;
     }
 
-    public int getCommission() {
+    public BigDecimal getCommission() {
         return commission;
     }
 
-    public void setCommission(int commission) {
+    public void setCommission(BigDecimal commission) {
         this.commission = commission;
     }
