+++ src/java/com/smi/travel/datalayer/report/model/OtherGuideCommissionSummaryHeader.java
@@ -5,6 +5,8 @@
  */
 package com.smi.travel.datalayer.report.model;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author Kanokporn
@@ -16,7 +18,7 @@ public class OtherGuideCommissionSummaryHeader {
     private String dateto;
     private String guidename;
     private int pax;
-    private int commission;
+    private BigDecimal commission;
 
     public String getSystemdate() {
         return systemdate;
@@ -66,12 +68,12 @@ public class OtherGuideCommissionSummaryHeader {
         this.pax = pax;
     }
 
-    public int getCommission() {
+    public BigDecimal getCommission() {
         return commission;
     }
 
-    public void setCommission(int commission) {
+    public void setCommission(BigDecimal commission) {
         this.commission = commission;
     }
-    
+   
 }
